fixed current location bug and set functionality of pin drop button
