+++ app/javascript/packs/application.js
@@ -3,18 +3,21 @@ import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the styleshe
 import '@mapbox/mapbox-gl-directions/dist/mapbox-gl-directions.css'
 
 import { initMapbox } from '../plugins/init_mapbox';
+import { otherInitMap } from '../plugins/map'
 // import { init_scroll } from '../plugins/init_scroll';
 
-// if (navigator.geolocation) {
-//   navigator.geolocation.getCurrentPosition(({coords: {latitude, longitude}}) => {
-//     console.log('load from current location');
-//     initMapbox([longitude, latitude]);
-//   });
-// } else {
-//   console.log('load from hardcoded location');
-//   initMapbox([115.1304015, -8.6539913]);
-// }
-initMapbox([115.1304015, -8.6539913]);
+if (navigator.geolocation) {
+  navigator.geolocation.getCurrentPosition(({coords: {latitude, longitude}}) => {
+    console.log('load from current location');
+    initMapbox([longitude, latitude]);
+    otherInitMap();
+  });
+} else {
+  console.log('load from hardcoded location');
+  initMapbox([115.1304015, -8.6539913]);
+  otherInitMap();
+}
+// initMapbox([115.1304015, -8.6539913]);
 // initScroll();
 
 if (navigator.serviceWorker) {
