+++ app/views/layouts/application.html.erb
@@ -16,6 +16,5 @@
     <%= yield %>
     <%= javascript_include_tag 'application' %>
     <%= javascript_pack_tag 'application' %>
-    <%= javascript_pack_tag 'map' %>
   </body>
 </html>
