+++ app/views/points/index.html.erb
@@ -10,7 +10,7 @@
 
   <footer id="map-footer">
     <form id="set-route" method="post" action="/trips" data-remote="true">
-      <input name="authenticity_token" id="auth" value="<%= form_authenticity_token %>"        type="hidden">
+      <input name="authenticity_token" id="auth" value="<%= form_authenticity_token %>" type="hidden">
       <input type="hidden" id="start_long" name="trip[start_long]" >
       <input type="hidden" id="start_lat" name="trip[start_lat]" >
       <input type="hidden" id="end_long" name="trip[end_long]" >
@@ -21,7 +21,8 @@
       </button>
     </form>
 
-    <%= link_to "", points_path, class: "fas fa-map-marker-alt fa-2x" %>
+    <button class="fas fa-map-marker-alt fa-2x" id="pin-drop"></button>
+    <!-- <%#= link_to "", points_path, class: "fas fa-map-marker-alt fa-2x" %> -->
 
     <%= link_to user_path(current_user[:id]) do %>
       <% avatar_url = current_user.facebook_picture_url || "#{current_user.avatar}" %>
